main_code/FGR/requirements.txt
# --------- pytorch --------- #
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4
captum
pytorch_multilabel_balanced_sampler
shap

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
hydra-joblib-launcher==1.2.0

# --------- loggers --------- #
wandb
dvc
dvc-gdrive

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh            # for running bash commands in some tests (linux/macos only)
omegaconf
tqdm
pynvml
pickleshare

# --------- chemistry --------- #
molvs
transformers
selfies
PyTDC
rdkit
CIRpy

# --------- dataframe manipulation --------- #
pandarallel
grpcio
pyarrow
openpyxl
h5py

# --------- visualization --------- #
seaborn
scikit_learn
matplotlib
jupyterlab
ipywidgets
# SciencePlots @ git+https://github.com/garrettj403/SciencePlots
umap-learn
